put global variables into EpochSys class
